+++ packages/create-gatsby/src/index.ts
@@ -15,12 +15,14 @@ import { trackCli } from "./tracking"
 import crypto from "crypto"
 import { reporter } from "./reporter"
 import { setSiteMetadata } from "./site-metadata"
+import { makeNpmSafe } from "./utils"
 
 const sha256 = (str: string): string =>
   crypto.createHash(`sha256`).update(str).digest(`hex`)
 
 const md5 = (str: string): string =>
   crypto.createHash(`md5`).update(str).digest(`hex`)
+
 /**
  * Hide string on windows (for emojis)
  */
@@ -52,6 +54,9 @@ const makeChoices = (
 export const validateProjectName = async (
   value: string
 ): Promise<string | boolean> => {
+  if (!value) {
+    return `You have not provided a directory name for your site. Please do so when running with the 'y' flag.`
+  }
   value = value.trim()
   if (INVALID_FILENAMES.test(value)) {
     return `The destination "${value}" is not a valid filename. Please try again, avoiding special characters.`
@@ -67,16 +72,17 @@ export const validateProjectName = async (
 
 // The enquirer types are not accurate
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
-export const questions: any = [
+export const questions = (initialFolderName: string, skip: boolean): any => [
   {
     type: `textinput`,
     name: `project`,
     message: `What would you like to name the folder where your site will be created?`,
     hint: path.basename(process.cwd()),
     separator: `/`,
-    initial: `my-gatsby-site`,
+    initial: initialFolderName,
     format: (value: string): string => c.cyan(value),
     validate: validateProjectName,
+    skip,
   },
   {
     type: `selectinput`,
@@ -101,6 +107,7 @@ export const questions: any = [
   },
 ]
 interface IAnswers {
+  name: string
   project: string
   styling?: keyof typeof styles
   cms?: keyof typeof cmses
@@ -137,6 +144,13 @@ export type PluginConfigMap = Record<string, Record<string, unknown>>
 const removeKey = (plugin: string): string => plugin.split(`:`)[0]
 
 export async function run(): Promise<void> {
+  const [flag, siteDirectory] = process.argv.slice(2)
+
+  let yesFlag = false
+  if (flag === `-y`) {
+    yesFlag = true
+  }
+
   trackCli(`CREATE_GATSBY_START`)
 
   const { version } = require(`../package.json`)
@@ -153,22 +167,47 @@ ${center(c.blueBright.bold.underline(`Welcome to Gatsby!`))}
 `
   )
 
-  reporter.info(
-    wrap(
-      `This command will generate a new Gatsby site for you in ${c.bold(
-        process.cwd()
-      )} with the setup you select. ${c.white.bold(
-        `Let's answer some questions:\n\n`
-      )}`,
-      process.stdout.columns
+  if (!yesFlag) {
+    reporter.info(
+      wrap(
+        `This command will generate a new Gatsby site for you in ${c.bold(
+          process.cwd()
+        )} with the setup you select. ${c.white.bold(
+          `Let's answer some questions:\n\n`
+        )}`,
+        process.stdout.columns
+      )
     )
-  )
+  }
 
   const enquirer = new Enquirer<IAnswers>()
 
   enquirer.use(plugin)
 
-  const data = await enquirer.prompt(questions)
+  let data
+  let siteName
+  if (!yesFlag) {
+    ;({ name: siteName } = await enquirer.prompt({
+      type: `textinput`,
+      name: `name`,
+      message: `What would you like to call your site?`,
+      initial: `My Gatsby Site`,
+      format: (value: string): string => c.cyan(value),
+    } as any))
+
+    data = await enquirer.prompt(questions(makeNpmSafe(siteName), yesFlag))
+  } else {
+    const warn = await validateProjectName(siteDirectory)
+    if (typeof warn === `string`) {
+      reporter.warn(warn)
+      return
+    }
+    siteName = siteDirectory
+    data = await enquirer.prompt(
+      questions(makeNpmSafe(siteDirectory), yesFlag)[0]
+    )
+  }
+
   data.project = data.project.trim()
 
   trackCli(`CREATE_GATSBY_SELECT_OPTION`, {
@@ -275,30 +314,36 @@ ${center(c.blueBright.bold.underline(`Welcome to Gatsby!`))}
 
     trackCli(`CREATE_GATSBY_SET_PLUGINS_STOP`)
   }
-
-  reporter.info(`
+  if (!yesFlag) {
+    reporter.info(`
 
 ${c.bold(`Thanks! Here's what we'll now do:`)}
 
     ${messages.join(`\n    `)}
   `)
 
-  const { confirm } = await new Enquirer<{ confirm: boolean }>().prompt({
-    type: `confirm`,
-    name: `confirm`,
-    initial: `Yes`,
-    message: `Shall we do this?`,
-    format: value => (value ? c.greenBright(`Yes`) : c.red(`No`)),
-  })
+    const { confirm } = await new Enquirer<{ confirm: boolean }>().prompt({
+      type: `confirm`,
+      name: `confirm`,
+      initial: `Yes`,
+      message: `Shall we do this?`,
+      format: value => (value ? c.greenBright(`Yes`) : c.red(`No`)),
+    })
 
-  if (!confirm) {
-    trackCli(`CREATE_GATSBY_CANCEL`)
+    if (!confirm) {
+      trackCli(`CREATE_GATSBY_CANCEL`)
 
-    reporter.info(`OK, bye!`)
-    return
+      reporter.info(`OK, bye!`)
+      return
+    }
   }
 
-  await initStarter(DEFAULT_STARTER, data.project, packages.map(removeKey))
+  await initStarter(
+    DEFAULT_STARTER,
+    data.project,
+    packages.map(removeKey),
+    siteName
+  )
 
   reporter.success(`Created site in ${c.green(data.project)}`)
 
@@ -308,7 +353,7 @@ ${c.bold(`Thanks! Here's what we'll now do:`)}
     reporter.info(`${w(`🔌 `)}Setting-up plugins...`)
     await installPlugins(plugins, pluginConfig, fullPath, [])
   }
-  await setSiteMetadata(fullPath, `title`, data.project)
+  await setSiteMetadata(fullPath, `title`, siteName)
 
   await gitSetup(data.project)
 
@@ -318,7 +363,7 @@ ${c.bold(`Thanks! Here's what we'll now do:`)}
   reporter.info(
     stripIndent`
     ${w(`🎉  `)}Your new Gatsby site ${c.bold(
-      data.project
+      siteName
     )} has been successfully created
     at ${c.bold(fullPath)}.
     `