+++ packages/gatsby-plugin-image/src/components/static-image.tsx
@@ -2,14 +2,22 @@ import {
   GatsbyImage as GatsbyImageBrowser,
   IGatsbyImageData,
 } from "./gatsby-image.browser"
-import { _getStaticImage } from "./static-image.server"
-import { StaticImageProps } from "../utils"
+import {
+  _getStaticImage,
+  propTypes,
+  IStaticImageProps,
+} from "./static-image.server"
 // These values are added by Babel. Do not add them manually
 interface IPrivateProps {
   __imageData?: IGatsbyImageData
   __error?: string
 }
 
-export const StaticImage: React.FC<
-  StaticImageProps & IPrivateProps
+const StaticImage: React.FC<
+  IStaticImageProps & IPrivateProps
 > = _getStaticImage(GatsbyImageBrowser)
+
+StaticImage.displayName = `StaticImage`
+StaticImage.propTypes = propTypes
+
+export { StaticImage }