+++ packages/gatsby-plugin-image/src/image-utils.ts
@@ -9,6 +9,7 @@ const DEFAULT_FIXED_WIDTH = 400
 export type Fit = "cover" | "fill" | "inside" | "outside" | "contain"
 
 export type Layout = "fixed" | "fluid" | "constrained"
+export type ImageFormat = "jpg" | "png" | "webp" | "avif" | "auto" | ""
 
 /**
  * The minimal required reporter, as we don't want to import it from gatsby-cli
@@ -65,8 +66,6 @@ export interface IImage {
   height: number
   format: ImageFormat
 }
-
-export type ImageFormat = "jpg" | "png" | "webp" | "avif" | "auto" | ""
 
 export interface IGatsbyImageHelperArgs {
   pluginName: string
@@ -143,7 +142,7 @@ export function generateImageData(
     pluginName,
     sourceMetadata,
     generateImageSource,
-    layout = `fixed`,
+    layout = `constrained`,
     fit,
     options,
     width,
@@ -280,7 +279,7 @@ export function calculateImageSizes(args: IImageSizeArgs): IImageSizes {
     height,
     maxHeight,
     filename,
-    layout = `fixed`,
+    layout = `constrained`,
     sourceMetadata: imgDimensions,
     reporter = { warn },
   } = args