+++ packages/gatsby/src/utils/handle-flags.ts
@@ -1,22 +1,14 @@
 import _ from "lodash"
 import { isCI } from "gatsby-core-utils"
-import realTerminalLink from "terminal-link"
+import terminalLink from "terminal-link"
 import { IFlag } from "./flags"
 import chalk from "chalk"
 import { commaListsAnd } from "common-tags"
 import { distance } from "fastest-levenshtein"
 
-const terminalLink = (text, url): string => {
-  if (process.env.NODE_ENV === `test`) {
-    return `${text} (${url})`
-  } else {
-    return realTerminalLink(text, url)
-  }
-}
-
 const handleFlags = (
   flags: Array<IFlag>,
-  configFlags: Record<string, boolean>,
+  configFlags: Record<string, boolean> = {},
   executingCommand = process.env.gatsby_executing_command
 ): {
   enabledConfigFlags: Array<IFlag>
@@ -25,11 +17,13 @@ const handleFlags = (
 } => {
   // Prepare config flags.
   // Filter out any flags that are set to false.
-  const availableFlags = new Map()
-  flags.forEach(flag => availableFlags.set(flag.name, flag))
+  const availableFlags = new Map<string, IFlag>()
+  flags.forEach(flag => {
+    availableFlags.set(flag.name, flag)
+  })
 
   // Find unknown flags someone has in their config to warn them about.
-  const unknownConfigFlags: Array<any> = []
+  const unknownConfigFlags: Array<{ flag: string; didYouMean: string }> = []
   for (const flagName in configFlags) {
     if (availableFlags.has(flagName)) {
       continue
@@ -60,25 +54,46 @@ const handleFlags = (
     unknownFlagMessage = commaListsAnd`The following flag(s) found in your gatsby-config.js are not known:`
     unknownConfigFlags.forEach(
       flag =>
-        (unknownFlagMessage += `\n- ${flag?.flag}${
-          flag?.didYouMean ? ` (did you mean: ${flag?.didYouMean})` : ``
+        (unknownFlagMessage += `\n- ${flag.flag}${
+          flag.didYouMean ? ` (did you mean: ${flag.didYouMean})` : ``
         }`)
     )
   }
 
-  let enabledConfigFlags = Object.keys(configFlags)
+  let enabledConfigFlags: Array<IFlag> = Object.keys(configFlags)
     .filter(name => configFlags[name] && availableFlags.has(name))
-    .map(flagName => availableFlags.get(flagName))
+    .map(flagName => availableFlags.get(flagName)!)
 
-  // If we're in CI, filter out any flags that don't want to be enabled in CI
-  if (isCI()) {
-    enabledConfigFlags = enabledConfigFlags.filter(flag => flag.noCi !== true)
-  }
+  // Test flags to see if it wants opted in.
+  const optedInFlags = new Map<string, IFlag>()
+  const applicableFlags = new Map<string, IFlag>()
+  availableFlags.forEach(flag => {
+    const fitness = flag.testFitness(flag)
 
-  // Filter out any flags that aren't for this environment.
-  enabledConfigFlags = enabledConfigFlags.filter(
-    flag => flag.command === `all` || flag.command === executingCommand
-  )
+    // if user didn't explicitly set a flag (either true or false)
+    // and it qualifies for auto opt-in - add it to optedInFlags
+    if (typeof configFlags[flag.name] === `undefined` && fitness === `OPT_IN`) {
+      enabledConfigFlags.push(flag)
+      optedInFlags.set(flag.name, flag)
+    }
+
+    if (fitness) {
+      applicableFlags.set(flag.name, flag)
+    }
+  })
+
+  // Filter enabledConfigFlags against various tests
+  enabledConfigFlags = enabledConfigFlags.filter(flag => {
+    // Is this flag available for this command?
+    const isForCommand =
+      flag.command === `all` || flag.command === executingCommand
+    // If we're in CI, filter out any flags that don't want to be enabled in CI
+    const isForCi = isCI() ? flag.noCI !== true : true
+
+    const passesFitness = flag.testFitness(flag)
+
+    return isForCommand && isForCi && passesFitness
+  })
 
   const addIncluded = (flag): void => {
     if (flag.includedFlags) {
@@ -118,13 +133,36 @@ const handleFlags = (
   let message = ``
   //  Create message about what flags are active.
   if (enabledConfigFlags.length > 0) {
-    message = `The following flags are active:`
-    enabledConfigFlags.forEach(flag => {
-      message += generateFlagLine(flag)
-    })
+    if (enabledConfigFlags.length - optedInFlags.size > 0) {
+      message = `The following flags are active:`
+      enabledConfigFlags.forEach(flag => {
+        if (!optedInFlags.has(flag.name)) {
+          message += generateFlagLine(flag)
+        }
+      })
+    }
 
-    const otherFlagsCount = flags.length - enabledConfigFlags.length
-    if (otherFlagsCount > 0) {
+    if (optedInFlags.size > 0) {
+      message += `\n\n`
+      message += `We're shipping new features! For final testing, we're rolling them out first to a small % of Gatsby users
+and your site was automatically chosen as one of them. With your help, we'll then release them to everyone in the next minor release.
+
+We greatly appreciate your help testing the change. Please report any feedback good or bad in the umbrella issue. If you do encounter problems, please disable the flag by setting it to false in your gatsby-config.js like:
+
+flags: {
+  THE_FLAG: false
+}
+
+The following flags were automatically enabled on your site:`
+      optedInFlags.forEach(flag => {
+        message += generateFlagLine(flag)
+      })
+    }
+
+    const otherFlagsCount = applicableFlags.size - enabledConfigFlags.length
+    // Check if there is other flags and if the user actually set any flags themselves.
+    // Don't count flags they were automatically opted into.
+    if (otherFlagsCount > 0 && Object.keys(configFlags).length > 0) {
       message += `\n\nThere ${
         otherFlagsCount === 1
           ? `is one other flag`
@@ -133,7 +171,7 @@ const handleFlags = (
 
       const enabledFlagsSet = new Set()
       enabledConfigFlags.forEach(f => enabledFlagsSet.add(f.name))
-      flags.forEach(flag => {
+      applicableFlags.forEach(flag => {
         if (!enabledFlagsSet.has(flag.name)) {
           message += generateFlagLine(flag)
         }